Rename crc.{h,cc} to crc_builder.{h,cc}.
Conflicts with CRC.h header on Mac, which has a case-insensitive file system.
