+++ main/autogen/BUILD
@@ -3,14 +3,14 @@ cc_library(
     srcs = [
         "autogen.cc",
         "autoloader.cc",
-        "crc.cc",
+        "crc_builder.cc",
         "packages.cc",
         "subclasses.cc",
     ],
     hdrs = [
         "autogen.h",
         "autoloader.h",
-        "crc.h",
+        "crc_builder.h",
         "packages.h",
         "subclasses.h",
     ],
